+++ src/Symfony/Component/PropertyInfo/Tests/Fixtures/DefaultGroupDummy.php
@@ -1,0 +1,17 @@
+<?php
+declare(strict_types=1);
+
+namespace Symfony\Component\PropertyInfo\Tests\Fixtures;
+
+use Symfony\Component\Serializer\Annotation\Groups;
+
+final class DefaultGroupDummy
+{
+
+    public $somethingWithoutGroup;
+
+    /**
+     * @Groups({"a"})
+     */
+    public $somethingWithGroup;
+}